webkit/data/layout_tests/platform/chromium-win-vista/LayoutTests/media/video-display-aspect-ratio-expected.txt

EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '426'), OBSERVED '320' FAIL
EXPECTED (video.videoHeight == '240') OK
END OF TEST
